requirements-dev.txt
Flask
requests
python-dotenv
gunicorn
PyJWT[crypto]
pytest
pytest-cov
flake8
mypy
bandit